#30 add exception handling to the character

+++ src/main/java/pl/piwniczneparagrafowki/thewagesofvoid/application/service/CharacterServiceImpl.java
@@ -1,15 +1,18 @@
 package pl.piwniczneparagrafowki.thewagesofvoid.application.service;
 
+import org.springframework.dao.DuplicateKeyException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import pl.piwniczneparagrafowki.thewagesofvoid.application.model.Character;
 import pl.piwniczneparagrafowki.thewagesofvoid.application.repository.CharacterRepository;
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
  * Class created by Arkadiusz Parafiniuk
- * arkadiusz.parafiniuk@gmail.com
+ * @Author arkadiusz.parafiniuk@gmail.com
  */
 @Service
 public class CharacterServiceImpl implements CharacterService {
@@ -21,6 +24,8 @@ public class CharacterServiceImpl implements CharacterService {
     public Character create(Character character) {
         if(characterRepository.findById(character.getId())==null) {
             characterRepository.save(character);
+        } else {
+            throw new DuplicateKeyException("CREATE FAILED: Object with id=" + character.getId() + " already exist in the database.");
         }
         return character;
     }
@@ -31,22 +36,40 @@ public class CharacterServiceImpl implements CharacterService {
         oldCharacter = characterRepository.findById(character.getId());
         if(oldCharacter!=null) {
             characterRepository.save(character);
+        } else {
+            throw new EmptyResultDataAccessException("UPDATE FAILED: Character with id=" + character.getId() + " not found in the database.", 1);
         }
         return character;
     }
 
     @Override
     public Character get(long id) {
-        return characterRepository.findById(id);
+        Character character;
+        character = characterRepository.findById(id);
+        if(character==null) {
+            throw new EmptyResultDataAccessException("GET FAILED: Character with id=" + id + " not found in the database.", 1);
+        }
+        return character;
     }
 
     @Override
     public List<Character> getAll() {
-        return characterRepository.findAll();
+        List<Character> characters = new ArrayList<>();
+        characters = characterRepository.findAll();
+        if(characters.isEmpty()) {
+            throw new EmptyResultDataAccessException("GET ALL FAILED: No characters found in the database.", 1);
+        }
+        return characters;
     }
 
     @Override
     public void delete(Character character) {
-        characterRepository.delete(character);
+        Character tmpCharacter;
+        tmpCharacter = characterRepository.findById(character.getId());
+        if(tmpCharacter!=null){
+            characterRepository.delete(character);
+        } else {
+            throw new EmptyResultDataAccessException("GET FAILED: Character with id=" + character.getId() + " not found in the database.", 1);
+        }
     }
 }
